DependencyAnalyzer: check known dependencies first
This fixes the byteCodeAnalyzer test, which failed nondeterministically
depending on the order files in a directory were iterated.

There is still an issue with the fact that dependencies are currently
recorded with version numbers, which might change later. But:

A) The plan is to switch to storing unversioned dependencies; and
B) The unit test does not have this issue, of course.

Thanks to Johannes Schindelin for his assistance.