getInstalledVersions: skip legacy entries
When we switched the checksumming to skipping certain parts of the
pom.xml and the manifest, we needed a way to represent different
checksums for the same file. These checksums were stored in the cached
checksums for speed.

However, almost two years after that switch, this information is no
longer useful for Fiji's bug submitter (for which we invented this
method).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>