UpdaterTestUtils: work around bug in scijava-common 2.20.0
There was a bug in 2.20.0 (which pom-imagej 2.15 references) that has been
fixed in the meantime. It only affected tests, so the bug fix is not
crucial enough to merit a full release cascade.

Let's have a work around just so we can run the tests even with the buggy
TestUtils class.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>