When deleting conflicting files, be extra careful on Windows
Windows. Yes, that Windows, this operating system that does everything
ever so slightly different, just enough to cost every cross-platform
developer a little bit of time every single day. And a bit of gray hair.

The problem is that the file we want to delete might be in use. Given
that it is on the class path, it is actually almost guaranteed to be
in use. And Windows, being special, at least in certain ways, cannot
delete files that are in use.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>