Introduce FileObject#isUploadable for files about to be modified
When users want to edit metadata in the updater GUI, we ask the file
object whether it can be uploaded. However, in this case the file's
metadata is about to be changed, but not yet changed. So we need a way
to ask: *could* this file be uploaded if the metadata were changed?

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>