Add another Micro-Manager specific directory
There is a very special setup required for Olympus microscopes, so
special indeed that a certain .dll has to be present in *two*
directories and it cannot simply be linked against, but the current
working directory has to be changed in order to use it which is why
Micro-Manager has to ship with a very specific executable in a
very specific place.

As Micro-Manager is an incredibly useful piece of software, and as
Olympus microscopes are pretty decent, at least hardware-wise, let's
add this very specific quirk to the updater just for Olympus.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>