Add the (hidden) 'history' function to the command-line updater
It is not really correct what this function shows because we have no
record *when* now-obsolete files became obsolete, or whether they
had been obsolete anytime in between. Nor, for that matter when files
got unshadowed/shadowed.

So let's keep this function out of the usage lest people expect it
to work flawlessly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>