Add a hidden command-line option for downgrading
Due to lack of metadata, the exact time when files became
obsolete/unshadowed is impossible to determine. As a consequence,
downgrading cannot be accurate. Therefore, we will keep this method
hidden because it is very likely that unsuspecting users, even when
they consider themselves savvy ones, actually in particular those,
will shoot themselves into the foot using this function.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>