UpdateService: add API to get a file's update site
This change adds a little bit of infrastructure to keep an internal
FilesCollection instance in the DefaultUpdateService, which is
initialized lazily. The FilesCollection is then used as needed
when relevant UpdateService methods are called.

The first such method is getUpdateSite(File), which will be needed
by the UsageUploadService when deciding which statistics are worthy
of being uploaded to the ImageJ usage server.