Introduce the 'revert-unreal-changes' command
When .dll files are recompiled, their checksums change. We would like
to revert those changes, and only those changes, when uploading, say,
the Micro-Manager nightly build.

This command handles .dll files for the moment, but the idea is to
be able to handle all kinds of 'unreal' file changes in the future,
via plugins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>