UpToDate: improve detection of developer setting
The previous detection was failing when launching net.imagej.Main from
Eclipse -- and probably in many other scenarios, too -- due to the fact
that we now use release couplings by default. So
net.imagej.updater.UpToDate.class _will_ be located in a JAR file, even
when running from the IDE. But the imagej.dir property is normally not
set in that scenario, since only the ImageJ Launcher sets it by default.

Of course, you could set imagej.dir in your Run configuration, which
would defeat this heuristic. But this approach should work for most
people most of the time...