HobbesUploader: work around bugs in SJC
Specifically, version 2.35.0 of scijava-common has a bug where
getTitle() returns null if the plugin is Identifiable but the class
cannot be loaded for some reason (i.e., because loadClass() returns
null). That bug is fixed with the following commit:

https://github.com/scijava/scijava-common/commit/93c340f472401567dee50f2edbb70abbff261749

However, to avoid the test failure before that fix gets released,
let's avoid the issue by giving the HobbesUploader a specific name,
which will be returned by getTitle() before the identifier is even
checked.

As an aside: while getTitle() is never supposed to return null, it still
wouldn't be an issue if the AbstractUIDetails#compareTo method were more
tolerant of null titles. That fix was also committed:

https://github.com/scijava/scijava-common/commit/bce7b69391791f39b5019dca14bf110816c97265