PromptUserToUpdate: do not update while headless
This fix is in response to a bug reported on the ImageJ mailing list:
  https://list.nih.gov/cgi-bin/wa.exe?A2=IMAGEJ;221b82b2.1502

This fix is a stopgap to avoid the problem, but architecturally, we
probably should someday improve the approach of UpdaterUI plugins.
Currently, they are not tied in to UserInterface plugins at all, but
they should be, so that an UpdaterUI plugin is only activated when its
associated UI is activated. This is same strategy used by the
WidgetService for deciding whether a given InputWidget should be used.

So really, we should generalize this logic to a common UIPlugin
interface and (maybe) base class, so that it can be reused. See:

    https://github.com/scijava/scijava-common/issues/147