UpdateService: auto-check only once per day
This is distinct from the UpToDate code's latestNag, because that code
clears the nag when the software is found to be fully up to date.

We really want this automatic "soft check" to only run occasionally,
not every time, regardless of the result of the check.

This hopefully helps prevent inadvertent DOS attacks on imagej.net
infrastructure when running ImageJ on a cluster.