Put files in update folder after adding a site
When adding/enabling an update site from the Command Line updater via
`add-update-site`, we now proceed enough down the update so that the
relevant changes are staged in the "update" folder.

This creates some duplicated effort if the update is applied with the
`update` directive, but has the benefit of allowing the update to be
completed just by the "normal" method of starting ImageJ/Fiji - and is
generally clearer that an update is pending.

See #115