Archive the Fiji.app folder if any file changes
Because the app folder is signed as a whole, if any file changes the
signing will have to change. This also means that if any file in
Fiji.app is updated, we do not want to leave a .old version in the app
folder as that will invalidate the signing.

Instead, if we now detect any file in the .app folder is updated, we
archive the complete state of the app folder to a single .old.app, and
flag all files in the .app folder for installation/update.