Remove Dropbox mapping logic
Dropbox is not a reliable long-term way to serve an update site,
as evidenced by the fact that this logic was ever necessary at all,
due to changes made over time by the Dropbox technology. To encourage
reproducibility, and avoid building kingdoms in other people's castles,
we remove this logic in favor of encouraging use of sites.imagej.net.

As this removal breaks API compatibility, we bump the major digit.