Fix launcher platform mapping for macOS
TL;DR: we shall ship all macOS-based launcher files to all flavors of
macOS. The platform `macosx` shall now mean *any* flavor of macOS.

While some files (e.g. skinny shared libraries or JARs wrapping them)
will be labeled with platform macos-arm64 or macos64, some other files
must be shipped to both macOS variants. In particular, everything in
Fiji.app *must* be shipped together, or else the immutability of the
code-signed bundle is ruined. For consistency, even though the stable
outer Fiji.app is not code-signed, we also mark all files under Contents
as macosx, so that all macOS systems receive all of those files.

It might seem like this change is a gratuitous breakage of backward
compatibility. But the hope is that it will be more backward compatible
with existing db.xml.gz files that label macOS-specific files with
platform `macosx`. Now that we have x64 and arm64 flavors of macOS,
skinny binaries for either of those each have their own dedicated short
platform name, macos64 and macos-arm64 respectively, named consistently
with the Linux and Windows platforms, and allowing each file to be
tagged as a specific macOS flavor as appropriate.