Broaden the definition of "active platform"
It now includes special platform-specific directories such as
jars/macos-arm64. This is especially important on macOS, where every
launcher belongs to the architecture-agnostic `macosx` platform group,
rather than to an architecture-specific macOS platform. We want to
activate macos-arm64 on Apple Silicon Macs, and macos64 on Intel Macs,
whenever any of their files are present.

But with the previous logic, due to the lack of architecture-specific
launchers, each would only activate when it is current platform, meaning
it was not possible to activate them on non-matching platforms. In
particular, this limitation made it annoying to activate macos64 on an
Apple Silicon Mac, even though having macos64 is useful for running in
emulation mode with Rosetta to work around any macos-arm64 native
libraries not yet being available.